requirements.txt
tensorboard
git+https://github.com/openai/CLIP.git
scipy
matplotlib
natsort
tqdm
opencv-python
yacs
tabulate
termcolor
git+https://github.com/mattloper/chumpy.git
mediapipe
func_timeout
ultralytics
gdown
setuptools
pytorch-lightning
pyrender
